增加dsp相同自定义类型

+++ AFWDPP.SF/App.cpp
@@ -1,7 +1,7 @@
 #include "App.h"
 
 
-int add(int a, int b) {
-	int c = a - b;
-	return a + b+100;
+Uint8 add(Uint8 a, Uint8 b) {
+	Uint8 c = a - b;
+	return a + b + 100;
 }
